requirements.txt
Pillow>=5.4.1
pycairo>=1.16.2
